LayoutTests/platform/ios-simulator/fast/events/mouse-click-events-expected.txt

This tests what mouse events we send.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Left Mouse Button
FAIL eventLog should be mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) dblclick(0) . Was mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) .
Middle Mouse Button
FAIL eventLog should be mousedown(1) mouseup(1) click(1) mousedown(1) mouseup(1) click(1) dblclick(1) . Was mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) .
Right Mouse Button
FAIL eventLog should be mousedown(2) mouseup(2) mousedown(2) mouseup(2) . Was mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) .
4th Mouse Button
FAIL eventLog should be mousedown(1) mouseup(1) click(1) mousedown(1) mouseup(1) click(1) dblclick(1) . Was mousedown(0) mouseup(0) click(0) mousedown(0) mouseup(0) click(0) .
PASS successfullyParsed is true

TEST COMPLETE
